clean up code v1.0
