+++ client/src/App.js
@@ -1,4 +1,3 @@
-//import UIContainer from "./components/UIContainer";
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Header from './components/Header'
