+++ client/src/components/Card.js
@@ -1,242 +1,127 @@
 import React, { useState } from 'react'
 import '../css/card.css'
-import { Container, Row, } from 'react-bootstrap';
-//import m01 from '../images/m01.jpg'
-import CardTemplate from './CardTemplate'
-//import axios from 'axios'
-// import { createResults } from '../utils/API';
+import m01 from '../images/m01.jpg'
+import { Col, Modal } from 'react-bootstrap'
+import { createResults } from '../utils/API';
 
+export default function CardTemplate ({card, selected, setSelected}) {
+    //console.log(card, `this is card from card template`)
+    //const [selected, setSelected] = useState([])
+    const [showModal, setShowModal] = useState(false);
+    const [flip, setFlip] = useState(false)
 
-export default function Card ({ cards }){
-    const [selected, setSelected] = useState([])
-    //const [showModal, setShowModal] = useState(false);
-    //const [cardData, setCardData] = useState([])
-    // const [flip, setFlip] = useState({
-    //     "1": false,
-    //     "2": false,
-    //     "3": false,
-    //     "4": false,
-    //     "5": false,
-    //     "6": false,
-    //     "7": false,
-    //     "8": false,
-    //     "9": false,
-    //     "10": false,
-    //     "11": false,
-    //     "12": false,
-    //     "13": false,
-    //     "14": false,
-    //     "15": false,
-    //     "16": false,
-    //     "17": false,
-    //     "18": false,
-    //     "19": false,
-    //     "20": false,
-    //     "21": false,
-    //     "22": false,
-    //     "23": false,
-    //     "24": false,
-    //     "25": false,
-    //     "26": false,
-    //     "27": false,
-    //     "28": false,
-    //     "29": false,
-    //     "30": false,
-    //     "31": false,
-    //     "32": false,
-    //     "33": false,
-    //     "34": false,
-    //     "35": false,
-    //     "36": false,
-    //     "37": false,
-    //     "38": false,
-    //     "39": false,
-    //     "40": false,
-    //     "41": false,
-    //     "42": false,
-    //     "43": false,
-    //     "44": false,
-    //     "45": false,
-    //     "46": false,
-    //     "47": false,
-    //     "48": false,
-    //     "49": false,
-    //     "50": false,
-    //     "51": false,
-    //     "52": false,
-    //     "53": false,
-    //     "54": false,
-    //     "55": false,
-    //     "56": false,
-    //     "57": false,
-    //     "58": false,
-    //     "59": false,
-    //     "60": false,
-    //     "61": false,
-    //     "62": false,
-    //     "63": false,
-    //     "64": false,
-    //     "65": false,
-    //     "66": false,
-    //     "67": false,
-    //     "68": false,
-    //     "69": false,
-    //     "70": false,
-    //     "71": false,
-    //     "72": false,
-    //     "73": false,
-    //     "74": false,
-    //     "75": false,
-    //     "76": false,
-    //     "77": false,
-    //     "78": false,
-    // })
-
-    //console.log(selected)
-
-    // const handleSubmit = async (e) => {
-    //     e.preventDefault()
-    //     try {
+    const handleSubmit = async (e) => {
+        e.preventDefault()
+        try {
             
 
-    //         console.log(`I am submitting...`)
-    //         const token = localStorage.getItem('id_token')
-    //         if (!token){
-    //             alert(`Please sign in to get your results!`)
-    //             return;
-    //         }
-    //         const results = await createResults(token, selected)
-    //         console.log(results)
-    //         if (!results){
-    //             alert(`Results not found`)  
-    //         }
+            console.log(`I am submitting...`)
+            const token = localStorage.getItem('id_token')
+            if (!token){
+                alert(`Please sign in to get your results!`)
+                return;
+            }
+            const results = await createResults(token, selected)
+            console.log(results)
+            // if (!results){
+            //     alert(`Results not found`)  
+            // }
 
-    //         //SetFlip to false for all 
+            //SetFlip to false for all 
 
-    //         window.location.replace('/portal')
-    //     } catch (error) {
-    //         console.log(error)
-    //     }
+            window.location.replace('/portal')
+        } catch (error) {
+            console.log(error)
+        }
         
-    // }
+    }
 
-    // const handleFlip = async (e) => {
-    //     if (selected.length !== 3){
-    //         console.log(e)
-    //         const { target } = e;
-    //         const id = target.id;
-    //         const name = target.name;
-    //         if (!selected.includes(name)){
-    //             setSelected([...selected, name])
-    //             console.log(`Card ${id} was clicked`)
-    //             console.log(`Card ${name} was clicked`)
-    //             if (id){
-    //                 const newObj = {...flip}
-    //                 newObj[id] = true
-    //                 setFlip(newObj);
-    //             }
-    //         }
+    const handleFlip = async (e) => {
+        setFlip(!flip)
+        const example = [...selected, card]
+        setSelected(example)
+        console.log(example)
 
-    //     } else {
+        if(selected.length === 2){
+
+            setShowModal(true)
+            // const [ past, present , future ] = selected
+
+            // const token = localStorage.getItem('id_token')
+            // if (!token){
+            //     alert(`Please sign in to get your results!`)
+            //     return;
+            // }
+
+            // const resultOne = await createResults(token, past)
+            // const resultTwo = await createResults(token, present)
+            // const resultThree = await createResults(token, future)
+            // if (resultOne && resultTwo && resultThree){
+            //     setShowModal(true)
+            // }
             
-    //         console.log(selected, `this is selected`)
-    //         setShowModal(true)
-    //         // const [ nameOne, nameTwo, nameThree ] = selected
-    //         // try {
-    //         //     const cardOne = await axios.get(`api/cards/${nameOne}`)
-    //         //     const cardTwo = await axios.get(`api/cards/${nameTwo}`)
-    //         //     const cardThree = await axios.get(`api/cards/${nameThree}`)
-    //         //     console.log(cardOne, `this is card 1`)
-    //         //     console.log(cardTwo, `this is card 2`)
-    //         //     console.log(cardThree, `this is card 3`)
-    //         //     if (!cardOne && !cardTwo && !cardThree){
-    //         //         alert(`Opps! something went wrong`)
-    //         //     }
-    //         //     //setCardData(cardOne, cardTwo, cardThree)
-    //         //     const past = cardOne.data;
-    //         //     const present = await cardTwo.data;
-    //         //     const future = await cardThree.data;
-    //         //     console.log(past, `this is past`)
-    //         //     console.log(present, `this is present`)
-    //         //     console.log(future, `this is future`)
-                
-    //         //     setCardData(past, present, future)
-    //         //     console.log(cardData, `this is cardData`)
-    //         //     setShowModal(true)        
-                
-    //         // } catch (error) {
-    //         //     console.log(error)
-    //         // }
-            
-            
-    //     }
-        
-    // }
-    
+        }
+    }
     return (
-        <Container>
-            {/* <div>
-                <button className={submit ? 'btn btn-success mx-auto' : 'btn btn-success mx-auto disabled'}>
-                        Get Reading!
-                </button>
-            </div> */}
-            <Row >
+        <>
+        <Col md={2}>
+                        
+            <div className='card play-card'>
+                <div className={flip ? "card__inner is-flipped" : 'card__inner'} >
+                    <div className='card__face card__face--front'>
+                        <img id={card.id} name={card.name} src={m01} alt='card-front' className='card-fit' onClick={handleFlip}></img>
+                    </div>
+                    <div className='card__face card__face--back'>
+                        <h6 className='none'>{card.name}</h6>
+                        <img src={require(`../images/${card.img}`)} alt='card-back' className='card-fit'></img>
+                    </div>
+
+                </div>
+            </div>
+                        
+                        
+        </Col>
+        <Modal 
+        size='lg'
+        show={showModal}
+        onHide={() => setShowModal(false)}
+        aria-labelledby='cards-modal'>
+        <Modal.Header closeButton>
+            <Modal.Title id='cards-modal'>
+                <h3>Your Results are In!</h3>
+            </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+            <p>Get Your Full Reading by pressing 'Get Reading' Below</p>
+            <div className='row' >
                 
-                {cards.flatMap((card, index) => (
-                    <CardTemplate card={card} key={index} selected={selected} setSelected={setSelected}/>
-                    // <Col md={2} key={index}>
-                        
-                    //     <div className='card'>
-                    //         <div className={flip[`${card.id}`] ? "card__inner is-flipped" : 'card__inner'} >
-                    //             <div className='card__face card__face--front'>
-                    //                 <img id={card.id} name={card.name} src={m01} alt='card-front' className='card-fit' onClick={handleFlip}></img>
-                    //             </div>
-                    //             <div className='card__face card__face--back'>
-                    //                 <h6 className='none'>{card.name}</h6>
-                    //                 <img src={require(`../images/${card.img}`)} alt='card-back' className='card-fit'></img>
-                    //             </div>
+                {selected.map((data, index) => {
+                    return(
+                        <div key={index} className='col-md-4'>
+                            <p className='text-center'>{data.name}</p>
+                        <div className='img-adjust'>
+                            
+                            <img src={require(`../images/${data.img}`)} alt='data-img' className='img-fit'></img>
+                        </div>
+                        </div>
+                    )
+                })} 
+                
 
-                    //         </div>
-                    //     </div>
-                        
-                        
-                    // </Col>
-                ))}
-
-            </Row>
-            {/* <Modal 
-                size='lg'
-                show={showModal}
-                onHide={() => setShowModal(false)}
-                aria-labelledby='cards-modal'>
-                <Modal.Header closeButton>
-                    <Modal.Title id='cards-modal'>
-                        <h3>Your Results are In!</h3>
-                    </Modal.Title>
-                </Modal.Header>
-                <Modal.Body>
-                    <p>Get Your Full Reading by pressing 'Get Reading' Below</p>
-                    <div >
-                        {selected[0]}{selected[1]}{selected[2]}
-                        
-                        
-
-                    </div>
-                </Modal.Body>
-                <Modal.Footer>
-                    <div>
-                        <button className='btn btn-success ' onClick={handleSubmit}>
-                            Get Reading!
-                        </button>
-                        <button className='btn btn-danger ' handleModalClose={() => setShowModal(false)}>
-                            No thanks
-                        </button>
-                    </div>
-                    
-                </Modal.Footer>
-            </Modal> */}
+            </div>
+        </Modal.Body>
+        <Modal.Footer>
+            <div>
+                <button className='btn btn-success ' onClick={handleSubmit}>
+                    Get Reading!
+                </button>
+                <button className='btn btn-danger ' handleModalClose={() => setShowModal(false)}>
+                    No thanks
+                </button>
+            </div>
             
-        </Container>
-    
+        </Modal.Footer>
+    </Modal>
+    </>
     )
 }
